refactor: password recovery options and token management methods

+++ backend/services/core_services/main_services/main_auth_service.py
@@ -1,29 +1,26 @@
 import authorization as authorization_utils
 from services.core_services import MainServiceBase
-from services.postgres_service import Post, User
+from services.postgres_service import User
 from pydantic_schemas.pydantic_schemas_auth import *
 from services_types import JWTTypes
 
-from sqlalchemy.ext.asyncio import AsyncSession
-from fastapi import UploadFile
-from fastapi import HTTPException
 from uuid import uuid4
-import os
 
 from exceptions.exceptions_handler import web_exceptions_raiser, endpoint_exception_handler
 from exceptions.custom_exceptions import *
 
 class MainServiceAuth(MainServiceBase):
-    async def generate_set_of_tokens(self, user_id: str, email_confirmation_required: bool = False) -> RefreshAccessTokens:
+    async def generate_set_of_tokens(self, user_id: str, email_confirmation_required: bool = False, check_tokens_exist: bool = True) -> RefreshAccessTokens:
         potential_refresh_token = await self._RedisService.get_token_by_user_id(user_id=user_id, token_type="refresh")
         potential_access_token = await self._RedisService.get_token_by_user_id(user_id=user_id, token_type="acces")
 
-        if potential_access_token:
-            await self._RedisService.delete_jwt(jwt_token=potential_access_token, token_type="acces")
-        if potential_refresh_token:
-            await self._RedisService.delete_jwt(jwt_token=potential_refresh_token, token_type="refresh")
-        
-        return await self._JWT.generate_save_refresh_access_token(
+        if check_tokens_exist:
+            if potential_access_token:
+                await self._RedisService.delete_jwt(jwt_token=potential_access_token, token_type="acces")
+            if potential_refresh_token:
+                await self._RedisService.delete_jwt(jwt_token=potential_refresh_token, token_type="refresh")
+            
+        return await self._JWT.generate_save_set_of_refresh_access_tokens(
             user_id=user_id,
             email_confirmation_required=email_confirmation_required,
             redis=self._RedisService
@@ -47,7 +44,7 @@ class MainServiceAuth(MainServiceBase):
 
         if not await self._RedisService.check_jwt_existence(jwt_token=prepared_token, token_type=token_type):
             raise Unauthorized(detail=f"AuthService: User tried to authorize request by expired token: {token} token type: {token_type}", client_safe_detail="Invalid or expired token")
-        
+         
         if return_user:
             payload = self._JWT.extract_jwt_payload(jwt_token=prepared_token)
             user = await self._PostgresService.get_user_by_id(payload.user_id)
@@ -134,7 +131,7 @@ class MainServiceAuth(MainServiceBase):
         return await self._JWT.generate_save_token(user_id=user.user_id, redis=self._RedisService, token_type="password-recovery")
 
     @web_exceptions_raiser
-    async def change_password(self, user: User, credentials: ChangePasswordBody) -> None:
+    async def recover_password(self, user: User, credentials: PasswordRecoveryBody) -> None:
         """Actually changes password via token"""
         
         authorization_utils.validate_password(credentials.new_password)
@@ -143,10 +140,22 @@ class MainServiceAuth(MainServiceBase):
         user.password_hash = new_password_hash
 
     @web_exceptions_raiser
-    async def request_password_recovery(self, user: User, credentials: ChangePasswordBody) -> EmailToConfirm:
+    async def request_password_recovery(self, user: User, credentials: PasswordRecoveryBody) -> EmailToConfirm:
         """Requests password change, issues change password 2fa"""
         await self.__create_2fa(email=user.email, username=user.username)
         return EmailToConfirm(email_to_confirm=user.email)
+
+    @web_exceptions_raiser
+    async def change_password(self, user: User, credentials: ChangePasswordBody) -> RefreshAccessTokens:
+        authorization_utils.validate_password(credentials.new_password)
+
+        if not authorization_utils.check_password(credentials.old_password, user.password_hash):
+            raise Unauthorized(detail=f"AuthService: User: {user.user_id} tried to change password using wrong old password.", client_safe_detail="Old password didn't match")
+
+        user.password_hash = authorization_utils.hash_password(credentials.new_password)
+
+        await self._RedisService.deactivate_tokens_by_user_id(user.user_id)
+        return await self.generate_set_of_tokens(user.user_id, email_confirmation_required=False, check_tokens_exist=False)
 
     @web_exceptions_raiser
     async def issue_new_second_factor(self, email: str):
@@ -158,9 +167,13 @@ class MainServiceAuth(MainServiceBase):
         await self.__create_2fa(email=email, username=user.username)
 
     @web_exceptions_raiser
-    async def logout(self, tokens: RefreshAccessTokens) -> None:
+    async def logout_on_this_device(self, tokens: RefreshAccessTokens) -> None:
         await self._RedisService.delete_jwt(jwt_token=tokens.access_token, token_type="acces")
         await self._RedisService.delete_jwt(jwt_token=tokens.refresh_token, token_type="refresh")
+
+    @web_exceptions_raiser
+    async def logout_on_every_device(self, user: User) -> None:
+        await self._RedisService.deactivate_tokens_by_user_id(user_id=user.user_id)
 
     @web_exceptions_raiser
     async def refresh_token(self, refresh_token: str) -> AccessTokenSchema:
@@ -168,7 +181,6 @@ class MainServiceAuth(MainServiceBase):
         if not await self._RedisService.check_jwt_existence(jwt_token=prepared_token, token_type="refresh"):
             raise Unauthorized(detail=F"AuthService: User with refresh token: {refresh_token} that does not exist tried to refresh tokens.", client_safe_detail="Invalid or expired token")
         
-
         payload = self._JWT.extract_jwt_payload(jwt_token=prepared_token)
         user_id = payload.user_id
 
@@ -193,6 +205,6 @@ class MainServiceAuth(MainServiceBase):
             raise InvalidResourceProvided(detail=f"AuthService: User: {user.user_id} tried to delete his profile, but password didn't match.", client_safe_detail="Password didn't match")
         
         await self._PostgresService.delete_models_and_flush(user)
-        await self._RedisService.deactivate_tokens_by_id(user_id=user.user_id)
+        await self._RedisService.deactivate_tokens_by_user_id(user_id=user.user_id)
         await self._ImageStorage.delete_avatar_user(user_id=user.user_id)
